Moved interface into seperate model folder
