+++ src/app/components/detail-panel/detail-panel.component.ts
@@ -1,5 +1,7 @@
 import { Component, OnInit, ChangeDetectionStrategy, Input } from '@angular/core';
-import { Element, ElementType, MainService } from 'src/app/services/main.service';
+import { MainService } from 'src/app/services/main.service';
+import { ElementType } from 'src/app/model/element-type.model';
+import { Element } from 'src/app/model/element.model';
 
 @Component({
   selector: 'ui-detail-panel',
