+++ src/app/mockData/elementTypes.ts
@@ -1,4 +1,4 @@
-import { ElementType } from '../services/main.service';
+import { ElementType } from '../model/element-type.model';
 
 export const LocalElementTypes: Array<ElementType> = [
   {
