+++ src/app/model/element-type.model.ts
@@ -1,0 +1,4 @@
+export interface ElementType {
+  uri: string;
+  name: string;
+}
