+++ src/app/model/element.model.ts
@@ -1,0 +1,5 @@
+export interface Element {
+  uri: string;
+  type: string;
+  name: string;
+}
