+++ src/app/services/main.service.ts
@@ -2,18 +2,8 @@ import { Injectable } from '@angular/core';
 import { Observable, of } from 'rxjs';
 import { MyCollection } from '../mockData/elements';
 import { JapaneseElementTypes, LocalElementTypes } from '../mockData/elementTypes';
-
-export interface Element {
-  uri: string;
-  type: string;
-  name: string;
-}
-
-export interface ElementType {
-  uri: string;
-  name: string;
-}
-
+import { ElementType } from '../model/element-type.model';
+import { Element } from '../model/element.model';
 
 @Injectable({
   providedIn: 'root'
